wind

+++ Classes/QuestProtocals.hpp
@@ -221,7 +221,7 @@ public:
 class WindBarProtocal
 {
 public:
-    virtual void op_configWind(int wind) = 0; //设置风力数值，正表示往右（顺），负的表示往左（逆风）0无风
+    virtual void op_configWind(float wind) = 0; //设置风力数值，正表示往右（顺），负的表示往左（逆风）0无风
 };
 
 #endif /* QuestProtocals_hpp */
